requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Pillow==10.2.0
numpy==1.26.4
tensorflow==2.13.0
opencv-python==4.9.0.80
geocoder==1.38.1
requests==2.31.0
twilio==9.0.1
